Major rewrite of the API to support more complex decoration methods. Old API still available in legacy.lua
This major rewrite grew out of a desire to be able to have stalactites placed according to a global noise function, which meant the whole decoration approach had to be changed to allow for data to be efficiently carried over between nodes. In the process I cleaned up a lot of old inefficient code, some of it written back when I was still learning LUA in general.

The legacy.lua file contains the old deprecated code so in theory any mods depending on subterrane should still function as they did before.